Extract the MediaSource and AudioContext interop from `Clip`
This decouples the `Clip` entity from the Web Audio APIs. Now both
APIs share a common interface and the correct one is loaded
dynamically.
